refactor: Order 도메인의 productquantity를 Int에서 Long으로 변경한다

+++ src/test/kotlin/org/rooftop/order/domain/OrderFixture.kt
@@ -16,7 +16,7 @@ fun order(
 
 fun orderProduct(
     productId: Long = 2L,
-    productQuantity: Int = 100,
+    productQuantity: Long = 100,
     totalPrice: Long = 10_000_000,
 ): OrderProduct = OrderProduct(
     productId = productId,
